Backend_mails/requirements.txt
alembic==1.13.2
aniso8601==7.0.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-GraphQL==2.0.0
Flask-GraphQL-Auth==1.3.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-api-core==2.22.0
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.21.0
googleapis-common-protos==1.65.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphql-server-core==1.1.1
greenlet==3.1.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
oauthlib==3.2.2
packaging==24.1
promise==2.3
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.0.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
Rx==1.6.3
sentinel==1.0.0
singledispatch==3.7.0
six==1.16.0
sqlakeyset==2.0.1726021475
SQLAlchemy==2.0.36
strawberry-graphql==0.248.0
strawberry-sqlalchemy-mapper==0.4.5
strawberry_sqlalchemy==0.1.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.4
